dev: add jooq impl

+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/dto/response/GithubResponse.java
@@ -8,7 +8,7 @@ import java.time.OffsetDateTime;
 public record GithubResponse(@NotNull @JsonProperty(value = "full_name") String fullName,
                              @NotNull @JsonProperty(value = "pushed_at") OffsetDateTime pushedAt,
                              @NotNull @JsonProperty(value = "updated_at") OffsetDateTime updatedAt,
-                             @NotNull @JsonProperty(value = "open_issues_count") Long openIssuesCount,
-                             @NotNull @JsonProperty(value = "forks") Long forksCount) {
+                             @NotNull @JsonProperty(value = "open_issues_count") Integer openIssuesCount,
+                             @NotNull @JsonProperty(value = "forks") Integer forksCount) {
 
 }
